Fix missing features in previous commit
