+++ application/Common/SharedTest/TestHelpers/TestConsumer.cs
@@ -1,69 +1,12 @@
 ﻿using MORR.Shared.Events.Queue.Strategy;
+using SharedTest.TestHelpers;
 using System;
 using System.Diagnostics;
-using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+
 
 namespace SharedTest.Events.Queue.Strategy
 {
-    public class TestResult
-    {
-        private bool didComplete;
-        private Exception? exception;
-
-        public void Complete()
-        {
-            didComplete = true;
-        }
-
-        public void Fail(Exception exception)
-        {
-            didComplete = true;
-            this.exception = exception;
-        }
-
-        public void AssertDidCompleteSuccessfully()
-        {
-            Assert.IsTrue(didComplete);
-            Assert.IsNull(exception);
-        }
-
-        public void AssertDidCompleteThrowing<T>() where T : Exception
-        {
-            Assert.IsTrue(didComplete);
-            Assert.IsInstanceOfType(exception, typeof(T));
-        }
-
-        public void WasSuccess(ManualResetEvent resetEvent)
-        {
-            Debug.Assert(resetEvent != null);
-
-            if (didComplete && exception == null)
-            {
-                resetEvent.Set();
-            }
-            else
-            {
-                resetEvent.Reset();
-            }
-        }
-
-        public void DidFailThrowing<T>(ManualResetEvent resetEvent) where T : Exception
-        {
-            Debug.Assert(resetEvent != null);
-
-            if (exception?.GetType() == typeof(T))
-            {
-                resetEvent.Set();
-            }
-            else
-            {
-                resetEvent.Reset();
-            }
-        }
-    }
-
     /*
      * The TestConsumer class provides mock for a simple consumer which allows several validation
      * methods on consuming an event. Therefore the Consumer only consumes TestEvents.
@@ -82,6 +25,12 @@ namespace SharedTest.Events.Queue.Strategy
             this.strategy = strategy;
         }
 
+        /// <summary>
+        /// Starts consuming from the strategy.
+        /// </summary>
+        /// <param name="runsAsync">Defines whether the producing action should run asynchronously ('true') or synchronously ('false').</param>
+        /// <param name="continueCondition">The condition action for defining a producing completion.</param>
+        /// <param name="completionAction">The completion action is called on completion. It is also called if any exception occurred.</param>
         public void Consume(bool runsAsync, Func<TestEvent, int, bool> continueCondition, Action<TestResult> completionAction)
         {
             Debug.Assert(continueCondition != null);
